// ui/src/lib/api/generated/model/createPolicyInputType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Centoire API
 * Centoire community platform API
 * OpenAPI spec version: 0.1.0
 */

export type CreatePolicyInputType = typeof CreatePolicyInputType[keyof typeof CreatePolicyInputType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreatePolicyInputType = {
  user: 'user',
  source: 'source',
  all: 'all',
} as const;